--- src/restaurants/entities/restaurants.entities.ts
@@ -1,17 +1,0 @@
-import { Field, ObjectType } from "@nestjs/graphql";
-
-
-@ObjectType()
-export class Restaurant{
-    @Field(type=>String)
-    name: string;
-
-    @Field(type=>Boolean)
-    isVegan: boolean;
-
-    @Field(type=>String)
-    addess:string;
-
-    @Field(type=>String)
-    ownerName:string
-}
