+++ src/restaurants/restaurants.resolver.ts
@@ -1,6 +1,6 @@
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { createRestaurantDto } from './dtos/create-restaurant.dto';
-import { Restaurant } from './entities/restaurants.entities';
+import { Restaurant } from './entities/restaurant.entity';
 
 @Resolver(of => Restaurant)
 export class RestaurantResolver {
